NewTrip/requirements.txt
asgiref==3.5.2
awesome-slugify==1.6.5
Babel==2.10.3
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.1.3
django-debug-toolbar==3.5.0
django-ranged-response==0.2.0
gunicorn==20.1.0
idna==3.3
Pillow==9.1.1
psycopg2-binary==2.9.3
pytz==2022.1
regex==2022.6.2
sqlparse==0.4.2
tzdata==2022.1
Unidecode==0.4.21
urllib3==1.26.9
